+++ x/model/client/cli/tx_model.go
@@ -18,9 +18,9 @@ func CmdCreateModel() *cobra.Command {
 		partNumber                                 string
 		commissioningCustomFlow                    int32
 		commissioningCustomFlowUrl                 string
-		commissioningModeInitialStepsHint          uint64
+		commissioningModeInitialStepsHint          uint32
 		commissioningModeInitialStepsInstruction   string
-		commissioningModeSecondaryStepsHint        uint64
+		commissioningModeSecondaryStepsHint        uint32
 		commissioningModeSecondaryStepsInstruction string
 		userManualUrl                              string
 		supportUrl                                 string
@@ -81,7 +81,7 @@ the necessary details for how to configure the product for initial commissioning
 	cmd.Flags().StringVar(&commissioningCustomFlowUrl, FlagCommissioningCustomFlowUrl, "",
 		`commissioningCustomFlowURL SHALL identify a vendor specific commissioning URL for the 
 device model when the commissioningCustomFlow field is set to '2'`)
-	cmd.Flags().Uint64Var(&commissioningModeInitialStepsHint, FlagCommissioningModeInitialStepsHint, 0,
+	cmd.Flags().Uint32Var(&commissioningModeInitialStepsHint, FlagCommissioningModeInitialStepsHint, 0,
 		`commissioningModeInitialStepsHint SHALL 
 identify a hint for the steps that can be used to put into commissioning mode a device that 
 has not yet been commissioned. This field is a bitmap with values defined in the Pairing Hint Table. 
@@ -92,7 +92,7 @@ that a device that has not yet been commissioned will enter Commissioning Mode u
 values of commissioningModeSecondaryStepsHint. Certain values of CommissioningModeInitialStepsHint, 
 as defined in the Pairing Hint Table, indicate a Pairing Instruction (PI) dependency, and for these 
 values the commissioningModeInitialStepsInstruction SHALL be set`)
-	cmd.Flags().Uint64Var(&commissioningModeSecondaryStepsHint, FlagCommissioningModeSecondaryStepsHint, 0,
+	cmd.Flags().Uint32Var(&commissioningModeSecondaryStepsHint, FlagCommissioningModeSecondaryStepsHint, 0,
 		`commissioningModeSecondaryStepsHint SHALL identify a hint for steps that can 
 be used to put into commissioning mode a device that has already been commissioned. 
 This field is a bitmap with values defined in the Pairing Hint Table. For example, a value of 4 (bit 2 is set)
